# packages/api/openapi.yaml
openapi: 3.1.0
info:
  title: marimohub API
  version: 1.0.0
  description: API for managing Marimo notebooks, projects, and sandboxes
...
security:
  - cookieAuth: []
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: mh_session
      description: Session cookie minted by the OIDC login flow; browsers attach it
        automatically.
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          enum:
            - false
        error:
          type: object
          properties:
            code:
              type: string
              enum:
                - BAD_REQUEST
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - CONFLICT
                - GONE
                - PRECONDITION_FAILED
                - PAYLOAD_TOO_LARGE
                - VALIDATION_ERROR
                - RESOURCE_EXHAUSTED
                - NOT_INITIALIZED
                - SERVICE_UNAVAILABLE
                - INTERNAL_ERROR
            message:
              type: string
            request_id:
              type: string
          required:
            - code
            - message
      required:
        - success
        - error
...
    NotebookPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SnapshotNotebookEntry'
        next_cursor:
          type:
            - string
            - 'null'
      required:
        - items
        - next_cursor
...
    NotebookMeta:
      type: object
      properties:
        id:
          type: string
        project_id:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum:
            - draft
            - active
            - archived
            - deleted
        author:
          type: string
        created_at:
          type: string
          format: date-time
          example: 2025-03-05T14:00:00Z
        updated_at:
          type: string
          format: date-time
          example: 2025-03-05T14:00:00Z
        last_run_at:
          type:
            - string
            - 'null'
          format: date-time
          example: 2025-03-05T14:00:00Z
        tags:
          type: array
          items:
            type: string
        runtime:
          type: object
          properties:
            python_version:
              type: string
            marimo_version:
              type: string
      required:
        - id
        - project_id
        - title
        - description
        - status
        - author
        - created_at
        - updated_at
        - last_run_at
        - tags
    GitNotebookCreateResult:
      type: object
      properties:
        notebook:
          $ref: '#/components/schemas/NotebookMeta'
        sync_url:
          type: string
        sync_token:
          type: string
      required:
        - notebook
        - sync_url
        - sync_token
    SyncToken:
      type: object
      properties:
        sync_url:
          type: string
        sync_token:
          type: string
      required:
        - sync_url
        - sync_token
...
paths:
...
  /api/v1/projects/{pid}/notebooks:
    get:
      tags:
        - Notebooks
      summary: List notebooks in a project
      parameters:
        - schema:
            type: string
            pattern: ^proj-[0-9a-z]{16}$
            example: proj-7h2k9qm4xz7rp3w8
          required: true
          name: pid
          in: path
        - schema:
            type: integer
            exclusiveMinimum: 0
            example: 100
          required: false
          name: limit
          in: query
        - schema:
            type: string
            example: WyIyMDI1LTAzLTA1VDE0OjAwOjAwWiIsIm5iLTEiXQ
          required: false
          name: cursor
          in: query
      responses:
        '200':
          description: List of notebooks, newest first
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  data:
                    $ref: '#/components/schemas/NotebookPage'
                required:
                  - success
                  - data
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service unavailable
          headers:
            Retry-After:
              schema:
                type: string
                description: Seconds to wait before retrying.
                example: '5'
              required: true
              description: Seconds to wait before retrying.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Notebooks
      summary: Create a notebook
      parameters:
        - schema:
            type: string
            pattern: ^proj-[0-9a-z]{16}$
            example: proj-7h2k9qm4xz7rp3w8
          required: true
          name: pid
          in: path
        - schema:
            type: string
            minLength: 1
            maxLength: 255
            example: a1b2c3d4-e5f6-7890
          required: false
          name: idempotency-key
          in: header
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  minLength: 1
                  example: Revenue Analysis
                description:
                  type: string
                  example: Monthly revenue breakdown
                code:
                  type: string
                  example: import marimo as mo
                tags:
                  type: array
                  items:
                    type: string
                  example:
                    - finance
                readme:
                  type: string
                deps:
                  type: string
                runtime:
                  type: object
                  properties:
                    python_version:
                      type: string
                    marimo_version:
                      type: string
              required:
                - title
                - description
                - code
      responses:
        '201':
          description: Notebook created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  data:
                    $ref: '#/components/schemas/NotebookMeta'
                required:
                  - success
                  - data
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service unavailable
          headers:
            Retry-After:
              schema:
                type: string
                description: Seconds to wait before retrying.
                example: '5'
              required: true
              description: Seconds to wait before retrying.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/projects/{pid}/notebooks/git:
    post:
      tags:
        - Notebooks
      summary: Create a git-synced workspace notebook
      parameters:
        - schema:
            type: string
            pattern: ^proj-[0-9a-z]{16}$
            example: proj-7h2k9qm4xz7rp3w8
          required: true
          name: pid
          in: path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  minLength: 1
                  example: GitHub app
                description:
                  type: string
                  example: Synced from a git repository
                provider:
                  type: string
                  enum:
                    - github
                  example: github
                repo:
                  type: string
                  minLength: 1
                  example: marimo-team/marimohub
                branch:
                  type: string
                  minLength: 1
                  example: main
                root_path:
                  type: string
                  example: apps
                entry_notebook:
                  type: string
                  minLength: 1
                  example: my_app.py
                tags:
                  type: array
                  items:
                    type: string
                  example:
                    - git
                readme:
                  type: string
                runtime:
                  type: object
                  properties:
                    python_version:
                      type: string
                    marimo_version:
                      type: string
              required:
                - title
                - description
                - repo
                - branch
                - entry_notebook
      responses:
        '201':
          description: Git-synced notebook created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  data:
                    $ref: '#/components/schemas/GitNotebookCreateResult'
                required:
                  - success
                  - data
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service unavailable
          headers:
            Retry-After:
              schema:
                type: string
                description: Seconds to wait before retrying.
                example: '5'
              required: true
              description: Seconds to wait before retrying.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/projects/{pid}/notebooks/{nid}/sync-token/rotate:
    post:
      tags:
        - Notebooks
      summary: Rotate a notebook sync token
      parameters:
        - schema:
            type: string
            pattern: ^proj-[0-9a-z]{16}$
            example: proj-7h2k9qm4xz7rp3w8
          required: true
          name: pid
          in: path
        - schema:
            type: string
            pattern: ^nb-[0-9a-z]{16}$
            example: nb-3w8h2k9qm4xz7rp3
          required: true
          name: nid
          in: path
      responses:
        '200':
          description: Sync token rotated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  data:
                    $ref: '#/components/schemas/SyncToken'
                required:
                  - success
                  - data
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service unavailable
          headers:
            Retry-After:
              schema:
                type: string
                description: Seconds to wait before retrying.
                example: '5'
              required: true
              description: Seconds to wait before retrying.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/projects/{pid}/notebooks/{nid}/content:
    get:
      tags:
        - Notebooks
      summary: Get notebook code
      parameters:
        - schema:
            type: string
            pattern: ^proj-[0-9a-z]{16}$
            example: proj-7h2k9qm4xz7rp3w8
          required: true
          name: pid
          in: path
        - schema:
            type: string
            pattern: ^nb-[0-9a-z]{16}$
            example: nb-3w8h2k9qm4xz7rp3
          required: true
          name: nid
          in: path
      responses:
        '200':
          description: Notebook source code
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  data:
                    type: object
                    properties:
                      code:
                        type: string
                    required:
                      - code
                required:
                  - success
                  - data
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service unavailable
          headers:
            Retry-After:
              schema:
                type: string
                description: Seconds to wait before retrying.
                example: '5'
              required: true
              description: Seconds to wait before retrying.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...