# packages/api/src/routes/notebooks.ts
import { createRoute, z } from '@hono/zod-openapi';
import type { Context } from 'hono';
import { zipSync } from 'fflate';
import {
	NotebookId,
	NotFoundError,
	ProjectId,
	toPublicNotebookMeta,
	toPublicSource,
	toPublicVersion,
	VersionId,
} from '@marimo-hub/core';
import {
	assertProjectRole,
	assertProjectVisible,
	commonErrors,
	createApp,
	errorResponses,
	etagFor,
	EtagResponseHeader,
	fail,
	ifMatchToken,
	IfMatchHeader,
	IdempotencyKeyHeader,
	jsonBody,
	jsonContent,
	NotebookDetailResponseSchema,
	NotebookIdParam,
	NotebookMetaResponseSchema,
	ProjectIdParam,
	RuntimeResponseSchema,
	NotebookVersionResponseSchema,
	SnapshotNotebookEntrySchema,
	SuccessResponseSchema,
} from '../shared';
import { idempotentCreate } from '../idempotency';
import type { HonoEnv } from '../context';
import { pageSchema, paginate, PaginationQuery } from '../pagination';

// --- Request body schemas ---

const CreateNotebookBody = z.object({
	title: z.string().min(1).openapi({ example: 'Revenue Analysis' }),
	description: z.string().openapi({ example: 'Monthly revenue breakdown' }),
	code: z.string().openapi({ example: 'import marimo as mo' }),
	tags: z
		.array(z.string())
		.optional()
		.openapi({ example: ['finance'] }),
	readme: z.string().optional(),
	deps: z.string().optional(),
	runtime: RuntimeResponseSchema.optional(),
});

const CreateGitNotebookBody = z.object({
	title: z.string().min(1).openapi({ example: 'GitHub app' }),
	description: z.string().openapi({ example: 'Synced from a git repository' }),
	provider: z.literal('github').optional().openapi({ example: 'github' }),
	repo: z.string().min(1).openapi({ example: 'marimo-team/marimohub' }),
	branch: z.string().min(1).openapi({ example: 'main' }),
	// Repo subdirectory whose tree is mirrored; `entry_notebook` is relative to it.
	root_path: z.string().optional().openapi({ example: 'apps' }),
	entry_notebook: z.string().min(1).openapi({ example: 'my_app.py' }),
	tags: z
		.array(z.string())
		.optional()
		.openapi({ example: ['git'] }),
	readme: z.string().optional(),
	runtime: RuntimeResponseSchema.optional(),
});

const SyncTokenResponseSchema = z
	.object({
		sync_url: z.string(),
		sync_token: z.string(),
	})
	.openapi('SyncToken');

const GitNotebookCreateResponseSchema = z
	.object({
		notebook: NotebookMetaResponseSchema,
		sync_url: z.string(),
		sync_token: z.string(),
	})
	.openapi('GitNotebookCreateResult');

const UpdateNotebookBody = z.object({
	title: z.string().min(1).optional(),
	description: z.string().optional(),
	code: z.string().optional(),
	tags: z.array(z.string()).optional(),
	readme: z.string().optional(),
	deps: z.string().optional(),
	message: z.string().optional().openapi({ example: 'Add regional breakdown' }),
});

const DuplicateNotebookBody = z.object({
	title: z.string().min(1).optional().openapi({ example: 'Revenue Analysis (copy)' }),
});

const VersionIdParam = NotebookIdParam.extend({
	vid: z
		.string()
		.regex(/^ver_[0-9A-Z]{26}$/)
		.refine(VersionId.is)
		.openapi({ param: { name: 'vid', in: 'path' }, example: 'ver_01HXYZ33333RSTUVWXYZAB' }),
});

// --- Route definitions ---

const listNotebooks = createRoute({
	method: 'get',
	path: '/projects/{pid}/notebooks',
	tags: ['Notebooks'],
	summary: 'List notebooks in a project',
	request: { params: ProjectIdParam, query: PaginationQuery },
	responses: {
		200: jsonContent(
			z.object({
				success: z.literal(true),
				data: pageSchema(SnapshotNotebookEntrySchema, 'NotebookPage'),
			}),
			'List of notebooks, newest first',
		),
		...commonErrors(),
		...errorResponses(400, 404),
	},
});

const createNotebook = createRoute({
	method: 'post',
	path: '/projects/{pid}/notebooks',
	tags: ['Notebooks'],
	summary: 'Create a notebook',
	request: {
		params: ProjectIdParam,
		headers: IdempotencyKeyHeader,
		body: jsonBody(CreateNotebookBody),
	},
	responses: {
		201: jsonContent(
			z.object({ success: z.literal(true), data: NotebookMetaResponseSchema }),
			'Notebook created',
		),
		...commonErrors(),
		...errorResponses(403, 404),
	},
});

const createGitNotebook = createRoute({
	method: 'post',
	path: '/projects/{pid}/notebooks/git',
	tags: ['Notebooks'],
	summary: 'Create a git-synced workspace notebook',
	request: { params: ProjectIdParam, body: jsonBody(CreateGitNotebookBody) },
	responses: {
		201: jsonContent(
			z.object({ success: z.literal(true), data: GitNotebookCreateResponseSchema }),
			'Git-synced notebook created',
		),
		...commonErrors(),
		...errorResponses(400, 403, 404),
	},
});

const rotateSyncToken = createRoute({
	method: 'post',
	path: '/projects/{pid}/notebooks/{nid}/sync-token/rotate',
	tags: ['Notebooks'],
	summary: 'Rotate a notebook sync token',
	request: { params: NotebookIdParam },
	responses: {
		200: jsonContent(
			z.object({ success: z.literal(true), data: SyncTokenResponseSchema }),
			'Sync token rotated',
		),
		...commonErrors(),
		...errorResponses(400, 403, 404),
	},
});

const getNotebook = createRoute({
	method: 'get',
	path: '/projects/{pid}/notebooks/{nid}',
	tags: ['Notebooks'],
	summary: 'Get notebook metadata',
	request: { params: NotebookIdParam },
	responses: {
		200: jsonContent(
			z.object({ success: z.literal(true), data: NotebookDetailResponseSchema }),
			'Notebook detail',
			EtagResponseHeader,
		),
		...commonErrors(),
		...errorResponses(404),
	},
});

const getNotebookContent = createRoute({
	method: 'get',
	path: '/projects/{pid}/notebooks/{nid}/content',
	tags: ['Notebooks'],
	summary: 'Get notebook code',
	request: { params: NotebookIdParam },
	responses: {
		200: jsonContent(
			z.object({
				success: z.literal(true),
				data: z.object({ code: z.string() }),
			}),
			'Notebook source code',
		),
		...commonErrors(),
		...errorResponses(404),
	},
});

const updateNotebook = createRoute({
	method: 'patch',
	path: '/projects/{pid}/notebooks/{nid}',
	tags: ['Notebooks'],
	summary: 'Update a notebook',
	request: { params: NotebookIdParam, headers: IfMatchHeader, body: jsonBody(UpdateNotebookBody) },
	responses: {
		200: jsonContent(
			z.object({ success: z.literal(true), data: NotebookMetaResponseSchema }),
			'Notebook updated',
			EtagResponseHeader,
		),
		...commonErrors(),
		...errorResponses(403, 404, 412),
	},
});

const deleteNotebook = createRoute({
	method: 'delete',
	path: '/projects/{pid}/notebooks/{nid}',
	tags: ['Notebooks'],
	summary: 'Delete a notebook (soft-delete)',
	request: { params: NotebookIdParam, headers: IfMatchHeader },
	responses: {
		200: jsonContent(SuccessResponseSchema, 'Notebook deleted'),
		...commonErrors(),
		...errorResponses(403, 404, 412),
	},
});

const listVersions = createRoute({
	method: 'get',
	path: '/projects/{pid}/notebooks/{nid}/versions',
	tags: ['Notebooks'],
	summary: 'List notebook versions',
	request: { params: NotebookIdParam, query: PaginationQuery },
	responses: {
		200: jsonContent(
			z.object({
				success: z.literal(true),
				data: pageSchema(NotebookVersionResponseSchema, 'NotebookVersionPage'),
			}),
			'List of versions, newest first',
		),
		...commonErrors(),
		...errorResponses(400, 404),
	},
});

const getVersion = createRoute({
	method: 'get',
	path: '/projects/{pid}/notebooks/{nid}/versions/{vid}',
	tags: ['Notebooks'],
	summary: 'Get a specific version',
	request: { params: VersionIdParam },
	responses: {
		200: jsonContent(
			z.object({
				success: z.literal(true),
				data: z.object({ version: NotebookVersionResponseSchema, code: z.string() }),
			}),
			'Version details with code',
		),
		...commonErrors(),
		...errorResponses(404),
	},
});

const getNotebookHtml = createRoute({
	method: 'get',
	path: '/projects/{pid}/notebooks/{nid}/html',
	tags: ['Notebooks'],
	summary: "Latest HTML snapshot of the notebook's outputs",
	description:
		"Serves the newest version's HTML snapshot (captured best-effort at session teardown) " +
		'raw — the static outputs shown to viewers under MARIMOHUB_VIEWER_MODE=static. ' +
		'`X-Marimohub-Version-Id` / `X-Marimohub-Captured-At` identify the snapshot. ' +
		'404 with code `NO_HTML_SNAPSHOT` when no version has one.',
	request: { params: NotebookIdParam },
	responses: {
		200: {
			content: { 'text/html': { schema: z.string() } },
			description: 'The HTML snapshot, served sandboxed (CSP forces an opaque origin)',
		},
		...commonErrors(),
		...errorResponses(404),
	},
});

const restoreVersion = createRoute({
	method: 'post',
	path: '/projects/{pid}/notebooks/{nid}/versions/{vid}/restore',
	tags: ['Notebooks'],
	summary: 'Restore a version as a new save',
	request: { params: VersionIdParam },
	responses: {
		201: jsonContent(
			z.object({ success: z.literal(true), data: NotebookMetaResponseSchema }),
			'Version restored as a new save; returns the updated notebook',
		),
		...commonErrors(),
		...errorResponses(403, 404),
	},
});

const duplicateNotebook = createRoute({
	method: 'post',
	path: '/projects/{pid}/notebooks/{nid}/duplicate',
	tags: ['Notebooks'],
	summary: 'Duplicate a notebook',
	request: {
		params: NotebookIdParam,
		headers: IdempotencyKeyHeader,
		body: jsonBody(DuplicateNotebookBody),
	},
	responses: {
		201: jsonContent(
			z.object({ success: z.literal(true), data: NotebookMetaResponseSchema }),
			'Notebook duplicated; returns the new notebook',
		),
		...commonErrors(),
		...errorResponses(403, 404),
	},
});

// --- App ---

const app = createApp();

function syncUrl(c: Context<HonoEnv>, pid: string, nid: string) {
	return `${new URL(c.req.url).origin}/api/sync/git/v1/projects/${pid}/notebooks/${nid}`;
}

app.openapi(listNotebooks, async (c) => {
	const deps = c.get('deps');
	const { notebooks, projects } = deps.services;
	const user = c.get('user');
	const { pid } = c.req.valid('param');
	await assertProjectVisible(projects, pid, user.id, deps.policy.defaultRole);
	const all = await notebooks.listNotebooks(pid);
	const data = paginate(all, c.req.valid('query'), {
		key: (n) => n.created_at,
		tiebreak: (n) => n.id,
	});
	return c.json({ success: true, data }, 200);
});

app.openapi(createNotebook, async (c) => {
	const deps = c.get('deps');
	const { notebooks, projects } = deps.services;
	const user = c.get('user');
	const { pid } = c.req.valid('param');
	await assertProjectRole(projects, pid, user.id, 'editor', deps.policy.defaultRole);
	const body = c.req.valid('json');
	const data = await idempotentCreate(c, 'POST /projects/{pid}/notebooks', async () => {
		const meta = await notebooks.createNotebook(pid, body, user.id);
		return toPublicNotebookMeta(meta);
	});
	return c.json({ success: true, data }, 201);
});

app.openapi(createGitNotebook, async (c) => {
	const deps = c.get('deps');
	const { notebooks, projects } = deps.services;
	const user = c.get('user');
	const { pid } = c.req.valid('param');
	await assertProjectRole(projects, pid, user.id, 'editor', deps.policy.defaultRole);
	const body = c.req.valid('json');
	const { meta, sync_token } = await notebooks.synced.create(pid, body, user.id);
	return c.json(
		{
			success: true,
			data: {
				notebook: toPublicNotebookMeta(meta),
				sync_url: syncUrl(c, pid, meta.id),
				sync_token,
			},
		},
		201,
	);
});

app.openapi(rotateSyncToken, async (c) => {
	const deps = c.get('deps');
	const { notebooks, projects } = deps.services;
	const user = c.get('user');
	const { pid, nid } = c.req.valid('param');
	await assertProjectRole(projects, pid, user.id, 'editor', deps.policy.defaultRole);
	const { sync_token } = await notebooks.synced.rotateToken(pid, nid);
	return c.json({ success: true, data: { sync_url: syncUrl(c, pid, nid), sync_token } }, 200);
});

app.openapi(getNotebook, async (c) => {
	const deps = c.get('deps');
	const { notebooks, projects } = deps.services;
	const user = c.get('user');
	const { pid, nid } = c.req.valid('param');
	await assertProjectVisible(projects, pid, user.id, deps.policy.defaultRole);
	const detail = await notebooks.getNotebook(pid, nid);
	const data = {
		meta: toPublicNotebookMeta(detail.meta),
		readme: detail.readme,
		source: toPublicSource(detail.source),
	};
	c.header('ETag', etagFor(detail.meta.updated_at));
	return c.json({ success: true, data }, 200);
});

app.openapi(getNotebookContent, async (c) => {
	const deps = c.get('deps');
	const { notebooks, projects } = deps.services;
	const user = c.get('user');
	const { pid, nid } = c.req.valid('param');
	await assertProjectVisible(projects, pid, user.id, deps.policy.defaultRole);
	const code = await notebooks.getNotebookContent(pid, nid);
	return c.json({ success: true, data: { code } }, 200);
});

app.openapi(updateNotebook, async (c) => {
	const deps = c.get('deps');
	const { notebooks, projects } = deps.services;
	const user = c.get('user');
	const { pid, nid } = c.req.valid('param');
	await assertProjectRole(projects, pid, user.id, 'editor', deps.policy.defaultRole);
	const body = c.req.valid('json');
	const meta = await notebooks.updateNotebook(pid, nid, body, user.id, ifMatchToken(c));
	c.header('ETag', etagFor(meta.updated_at));
	return c.json({ success: true, data: toPublicNotebookMeta(meta) }, 200);
});

app.openapi(deleteNotebook, async (c) => {
	const deps = c.get('deps');
	const { notebooks, projects } = deps.services;
	const user = c.get('user');
	const { pid, nid } = c.req.valid('param');
	await assertProjectRole(projects, pid, user.id, 'editor', deps.policy.defaultRole);
	await notebooks.deleteNotebook(pid, nid, user.id, ifMatchToken(c));
	return c.json({ success: true }, 200);
});

app.openapi(listVersions, async (c) => {
	const deps = c.get('deps');
	const { notebooks, projects } = deps.services;
	const user = c.get('user');
	const { pid, nid } = c.req.valid('param');
	await assertProjectVisible(projects, pid, user.id, deps.policy.defaultRole);
	const all = await notebooks.listVersions(pid, nid);
	const page = paginate(all, c.req.valid('query'), {
		key: (v) => v.saved_at,
		tiebreak: (v) => v.version_id,
	});
	const data = { ...page, items: page.items.map(toPublicVersion) };
	return c.json({ success: true, data }, 200);
});

app.openapi(getVersion, async (c) => {
	const deps = c.get('deps');
	const { notebooks, projects } = deps.services;
	const user = c.get('user');
	const { pid, nid, vid } = c.req.valid('param');
	await assertProjectVisible(projects, pid, user.id, deps.policy.defaultRole);
	const { version, code } = await notebooks.getVersion(pid, nid, vid);
	return c.json({ success: true, data: { version: toPublicVersion(version), code } }, 200);
});

app.openapi(getNotebookHtml, async (c) => {
	const deps = c.get('deps');
	const { notebooks, projects } = deps.services;
	const user = c.get('user');
	const { pid, nid } = c.req.valid('param');
	// Read-only, gated like reading the notebook's code (viewer visibility).
	await assertProjectVisible(projects, pid, user.id, deps.policy.defaultRole);
	await notebooks.getNotebook(pid, nid);
	const snapshot = await notebooks.getLatestHtmlSnapshot(pid, nid);
	if (!snapshot) {
		return fail(c, 'NO_HTML_SNAPSHOT', 'No HTML snapshot has been captured yet', 404);
	}
	// The snapshot is user-generated HTML (marimo's export embeds scripts) and must
	// never execute same-origin with the app. CSP `sandbox` forces an opaque origin
	// even when the URL is opened directly, independent of the client's own
	// <iframe sandbox> — two independent containment layers.
	c.header('Content-Security-Policy', 'sandbox allow-scripts');
	c.header('X-Content-Type-Options', 'nosniff');
	// Authenticated, user-generated content: never disk-cached (a shared machine
	// must not serve a private notebook's outputs after logout).
	c.header('Cache-Control', 'private, no-store');
	c.header('X-Marimohub-Version-Id', snapshot.versionId);
	c.header('X-Marimohub-Captured-At', snapshot.capturedAt);
	return c.html(snapshot.html, 200);
});

app.openapi(restoreVersion, async (c) => {
	const deps = c.get('deps');
	const { notebooks, projects } = deps.services;
	const user = c.get('user');
	const { pid, nid, vid } = c.req.valid('param');
	await assertProjectRole(projects, pid, user.id, 'editor', deps.policy.defaultRole);
	const meta = await notebooks.restoreVersion(pid, nid, vid, user.id);
	return c.json({ success: true, data: toPublicNotebookMeta(meta) }, 201);
});

app.openapi(duplicateNotebook, async (c) => {
	const deps = c.get('deps');
	const { notebooks, projects } = deps.services;
	const user = c.get('user');
	const { pid, nid } = c.req.valid('param');
	await assertProjectRole(projects, pid, user.id, 'editor', deps.policy.defaultRole);
	const body = c.req.valid('json');
	const data = await idempotentCreate(
		c,
		'POST /projects/{pid}/notebooks/{nid}/duplicate',
		async () => {
			const meta = await notebooks.duplicateNotebook(pid, nid, user.id, body.title);
			return toPublicNotebookMeta(meta);
		},
	);
	return c.json({ success: true, data }, 201);
});

// GET .../workspace.zip — download the notebook's `workspace/` mirror as a zip.
// A plain Hono route (not `app.openapi`) because the body is binary, not the
// `{success,data}` JSON envelope — keeping it out of the OpenAPI doc avoids
// generating an unusable binary method in the typed client. Auth/CSRF/init
// middleware still apply (it is mounted under `/api`). Read-only, like the other
// notebook GETs, so gated only at `viewer` (visibility) like them.
//
// The whole workspace is read into memory and zipped synchronously (`zipSync`).
// Workspaces are small (notebook + data files), so this is fine; `zipSync` is
// preferred over the async `zip()` because the latter's Worker path is unreliable
// in the bundled, no-`node_modules` server image.
app.get('/projects/:pid/notebooks/:nid/workspace.zip', async (c) => {
	const deps = c.get('deps');
	const { notebooks, projects } = deps.services;
	const user = c.get('user');
	const pidRaw = c.req.param('pid');
	const nidRaw = c.req.param('nid');
	if (!ProjectId.is(pidRaw) || !NotebookId.is(nidRaw)) {
		throw new NotFoundError('Notebook not found');
	}
	await assertProjectVisible(projects, pidRaw, user.id, deps.policy.defaultRole);

	const files = await notebooks.listWorkspaceFiles(pidRaw, nidRaw);
	const zipped = zipSync(Object.fromEntries(files.map((f) => [f.path, f.bytes])));

	return new Response(zipped, {
		headers: {
			'content-type': 'application/zip',
			'content-disposition': 'attachment; filename="workspace.zip"',
		},
	});
});

export default app;